requirements.txt
boto3
streamlit
streamlit_lottie
python-dotenv